add check on the toggle

+++ simpleToggle/toggle.sh
@@ -5,10 +5,16 @@
 
 search="$1" #process to search for. By default, it takes in an argument by the user (ex: ./toggle.sh onboard). Change as needed
 pid=$(pidof $search) #used for the if else logic below
+which=$(which $search &>/dev/null) 
 
-#if else statement to determine whether to spawn process or destroy it
-if [ ! -z "$pid" ];then
-    kill "$pid"
+if [ ! -z "$which" ];then
+    #if else statement to determine whether to spawn process or destroy it
+    if [ ! -z "$pid" ];then
+        kill "$pid"
+    else
+        $search </dev/null &>/dev/null & #don't keep the program running in terminal
+    fi
 else
-    $search </dev/null &>/dev/null & #don't keep the program running in terminal
+    echo "No matching process found."
 fi
+
